added a confirmation, beforer logout

+++ src/app/header/header.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-header',
@@ -8,14 +9,16 @@ import { Component, OnInit } from '@angular/core';
 export class HeaderComponent implements OnInit {
   loggedIn: boolean;
 
-  constructor() { }
+  constructor(private router: Router) { }
 
   ngOnInit(): void {
     this.loggedIn = localStorage.getItem('login') === 'true';
   }
 
   onLogout() {
+    if (!(confirm('Wollen Sie sich wirklich ausloggen?'))) return false;
     localStorage.setItem('login', 'false');
+    this.router.navigate(['/logout']);
   }
 
 }
